+++ NEWS.md
@@ -5,6 +5,7 @@
 - add `Suggests: seriation` to illustrate CA re-ordering of rows/cols based on correspondence analysis
 - all .Rd files reformatted and examples extended.
 - the vignette on mosaic displays has been extensively revised with examples for square tables and permutation of row / column variables.
+- begin to classify datasets with content tags
 
 ## Version 0.8-1 (2022-04-22)